XplatUIX11.WorkingArea can segfault if the WM does not support _NET_WORKAREA
The result checking for the call to XGetWindowProperty for the _NET_WORKAREA
property in XplatUIX11.WorkingArea's getter is incorrect, which can lead to a
segfault. Whether the call succeeded or not is checked by "if ((long)nitems < 4
* current_desktop)". The problems is that the count of desktops starts by 0, so
`nitems` must have at least a value of "4 * (current_desktop + 1)" and not "4 *
current_desktop".