In order to help out QEMU in non-system-emulation mode (where no useful /proc/cpuinfo is provided), this branch adds a MAYBE assumption that the platform PROBABLY supports ARMv5, ARMv7, and Thumb instructions, if no evidence is provided by /proc/cpuinfo that it's even a valid ARM CPU.
However, where values are provided, whilst v5_supported and thumb_supported
are set to TRUE ot FALSE depending on the running system, v7_supported is
only ever set to TRUE, i.e. the only possible values are "TRUE" or "MAYBE",
meaning ARMv7 instructions are always generated.

This commit simply inserts an explicit "v7_supported = FALSE" after the
QEMU get-out clause, but before the check for v7, so the default status on
non-QEMU hosts is FALSE.

This fixes the problem of trying to use v7 instructions on v6, v5 and v4.