Remove monodoc-base trigger
This trigger updated the monodoc search index. It ended up calling
/usr/bin/monodoc, which is shipped in the monodoc-browser package
(source package mono-tools). The script attempted to check that
monodoc-browser was configured, but didn't get this right. This led to
numerous upgrade failures when Depends of monodoc-browser were not
satisfied when the trigger was invoked and calling monodoc to update the
search index bombed out due to this. It's more correct to just have
monodoc-browser ship this trigger itself.

LP: #972751