Tweak build system to check multiarch library folder for libX11.so
Mono's build system checks /usr/lib and /usr/lib64 to determine the
soname for libX11, and add it to /etc/mono/config - if it can't
find the library in those locations, it uses an unversioned library
in the config file. This patch forces the build system to check the
current build folder's multiarch library folder, e.g.
/usr/lib/x86_64-linux-gnu/